refactor: Move RunRecordQueryBO from enums to entity package
